update config imports
