+++ ferritin-featurizers/src/models/ligandmpnn/mod.rs
@@ -1,3 +1,4 @@
+pub mod configs;
 pub mod featurizer;
 pub mod model;
 mod proteinfeatures;
