+++ ferritin-featurizers/src/models/ligandmpnn/model.rs
@@ -1,4 +1,5 @@
 // use super::python_compat::{LigandMPNNData, LigandMPNNDataDict};
+use super::configs::{ModelType, ProteinMPNNConfig};
 use super::featurizer::ProteinFeatures;
 use super::proteinfeatures::ProteinFeaturesModel;
 use super::utilities::{cat_neighbors_nodes, gather_nodes};
@@ -256,60 +257,6 @@ impl DecLayer {
         };
 
         Ok(h_v)
-    }
-}
-
-#[derive(Clone, Debug)]
-pub enum PMPNNModelType {
-    LigandMPNN,
-    ProteinMPNN,
-    SolubleMPNN,
-}
-
-#[derive(Clone, Debug)]
-pub struct ProteinMPNNConfig {
-    atom_context_num: usize,
-    augment_eps: f32,
-    dropout_ratio: f32,
-    edge_features: i64,
-    hidden_dim: i64,
-    k_neighbors: i64,
-    ligand_mpnn_use_side_chain_context: bool,
-    model_type: PMPNNModelType,
-    node_features: i64,
-    num_decoder_layers: i64,
-    num_encoder_layers: i64,
-    num_letters: i64,
-    num_rbf: i64,
-    scale_factor: f64,
-    vocab: i64,
-}
-
-impl ProteinMPNNConfig {
-    pub fn proteinmpnn() -> Self {
-        Self {
-            atom_context_num: 0,
-            augment_eps: 0.0,
-            dropout_ratio: 0.1,
-            edge_features: 128,
-            hidden_dim: 128,
-            k_neighbors: 24,
-            ligand_mpnn_use_side_chain_context: false,
-            model_type: PMPNNModelType::ProteinMPNN,
-            node_features: 128,
-            num_decoder_layers: 3,
-            num_encoder_layers: 3,
-            num_letters: 48,
-            num_rbf: 16,
-            scale_factor: 30.0,
-            vocab: 48,
-        }
-    }
-    fn ligandmpnn() {
-        todo!()
-    }
-    fn membranempnn() {
-        todo!()
     }
 }
